Refactor Home component and BlogSection component
The Home component in page.tsx has been updated to be an asynchronous function. This change was made to handle asynchronous operations in the component, improving overall performance and user experience.

Additionally, the BlogSection component in BlogSection.tsx has been refactored to include an asynchronous function `getData()`. This function retrieves data from an API using the `axios` library and returns a Promise of an array of BlogPost objects. The retrieved data is then used to populate the component with the necessary information for rendering.

These changes help enhance the functionality and efficiency of the Home and BlogSection components.