feat: Modify BlogSection component to fetch and display blog posts
The BlogSection component has been updated to fetch and display blog posts from the API. The `getData` function now uses the `fetch` API instead of the `api` module to retrieve the blog post data. The `response` is parsed as JSON using the `json` method. The retrieved blog post data is then used to populate the `posts` array, which is returned as the result of the `getData` function.

In the BlogSection component, the commented code that used the `api` module has been replaced with the updated `getData` function. The main blog item and side blog items are now rendered using a mapping function on the `data` array. The `MainBlogItem` and `SideBlogItem` components receive the blog data as props and use it to render the appropriate content.

The `MainBlogItem` component has been updated to use the photo, author, category, date, and link properties from the blog data. The photo, author photo, and author name are now rendered dynamically using the data from the blog object. The date is formatted using the `moment` library.

The `SideBlogItem` component has also been updated to use the photo, category, date, and link properties from the blog data. The photo is now rendered dynamically using the data from the blog object. The date is formatted using the `moment` library.

These modifications allow the BlogSection component to fetch and display blog posts dynamically from the API.