Add mobile device detection for smooth scrolling
Introduced a new utility function `isMobileDevice` to detect if the user is using a mobile device. Altered the `SmoothScroll` component to only initialize the `Lenis` scrolling library if the user is not on a mobile device. This is to prevent unwanted scrolling behaviors on mobiles that can negatively affect user experience on these devices.