Refactor Hero component and HeroSlider component
The Hero component has been modified to include a new state variable `isMobile` and an effect hook to set its value using the `isMobileDevice()` utility from "@/utils/device". The `isMobile` variable is then passed as a prop to the HeroSlider component. Additionally, the HeroSlider component now accepts a new prop `isMobile` of type boolean. The Swiper component within HeroSlider is also modified to conditionally set the `draggable` prop based on the value of `isMobile`. Finally, some parts of the HeroSlider component, namely the SwiperSlides for banner two and banner three, are wrapped in a conditional rendering based on the value of `isMobile`.